+++ Sources/SpeziFHIR/FHIRAttachment/ContentExtractor.swift
@@ -6,9 +6,9 @@
 // SPDX-License-Identifier: MIT
 //
 
-// swiftlint:disable file_types_order
+#if canImport(UniformTypeIdentifiers)
 
-import PDFKit
+import Foundation
 import UniformTypeIdentifiers
 
 
@@ -26,47 +26,4 @@ protocol ContentExtractor {
     func extractContent(from data: Data) throws -> String
 }
 
-
-/// Extractor for plain text content types.
-struct TextContentExtractor: ContentExtractor {
-    func isCompatible(with contentType: UTType) -> Bool {
-        contentType.conforms(to: .text)
-    }
-    
-    func extractContent(from data: Data) throws -> String {
-        guard let string = String(data: data, encoding: .utf8) else {
-            throw FHIRAttachmentError.textDecodingFailed
-        }
-        return string
-    }
-}
-
-
-/// Extractor for PDF document content types.
-struct PDFContentExtractor<PDFProvider: PDFDocumentProviding>: ContentExtractor {
-    private let pdfDocumentProvider: PDFProvider
-    
-    /// Creates a new instance of the PDF content extractor.
-    /// - Parameter pdfDocumentProvider: The provider used to create PDFDocument instances.
-    init(pdfDocumentProvider: PDFProvider = DefaultPDFDocumentProvider()) {
-        self.pdfDocumentProvider = pdfDocumentProvider
-    }
-    
-    func isCompatible(with contentType: UTType) -> Bool {
-        contentType.conforms(to: .pdf)
-    }
-    
-    func extractContent(from data: Data) throws -> String {
-        guard let pdf = pdfDocumentProvider.createPDFDocument(from: data) else {
-            throw FHIRAttachmentError.pdfParsingFailed
-        }
-        let documentContent = NSMutableAttributedString()
-        for pageIdx in 0..<pdf.pageCount {
-            guard let page = pdf.page(at: pageIdx), let content = page.attributedString else {
-                continue
-            }
-            documentContent.append(content)
-        }
-        return documentContent.string
-    }
-}
+#endif
